+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add alertmanager ingress oauth endpoint.
 - Add resources for alertmanager ingress.
+- Wire prometheus to both alertmanagers.
 
 ### Changed